ProductsToPickRowsDataFactory: refactor it and break it in multiple small classes
https://github.com/metasfresh/metasfresh/issues/5679
