+++ src/main/java/de/metas/ui/web/pickingV2/productsToPick/ProductsToPickViewFactory.java
@@ -14,6 +14,9 @@ import de.metas.ui.web.pickingV2.productsToPick.process.ProductsToPick_MarkWillN
 import de.metas.ui.web.pickingV2.productsToPick.process.ProductsToPick_PickSelected;
 import de.metas.ui.web.pickingV2.productsToPick.process.ProductsToPick_Request4EyesReview;
 import de.metas.ui.web.pickingV2.productsToPick.process.ProductsToPick_SetPackingInstructions;
+import de.metas.ui.web.pickingV2.productsToPick.rows.ProductsToPickRow;
+import de.metas.ui.web.pickingV2.productsToPick.rows.ProductsToPickRowsData;
+import de.metas.ui.web.pickingV2.productsToPick.rows.ProductsToPickRowsRepository;
 import de.metas.ui.web.view.CreateViewRequest;
 import de.metas.ui.web.view.IViewFactory;
 import de.metas.ui.web.view.IViewsRepository;
