+++ src/main/java/de/metas/ui/web/pickingV2/productsToPick/process/ProductsToPick_MarkWillNotPickSelected.java
@@ -8,7 +8,7 @@ import de.metas.handlingunits.picking.PickingCandidateService;
 import de.metas.handlingunits.picking.candidate.commands.RejectPickingResult;
 import de.metas.handlingunits.picking.requests.RejectPickingRequest;
 import de.metas.process.ProcessPreconditionsResolution;
-import de.metas.ui.web.pickingV2.productsToPick.ProductsToPickRow;
+import de.metas.ui.web.pickingV2.productsToPick.rows.ProductsToPickRow;
 
 /*
  * #%L
